Divide components into container and presentational
The components that will communicate with the Redux store are categorized as containers
The components that will deal with presentation are categorized as presentational components. Like, App.js

Create Reducer for Books.

The book reducer handles the deleting and adding of book based on the action type. The immutability is achieved
through spread syntax. The book object is recieved in the payload of action object.