Add filter functionality
Added the functionality to filter the books based on categories. To
implemenet the feature, I have create filter action with type
"CHANGE_FILTER". I have added a filter reducer that defaults to 'All'.
The reducer checks the action.filter and changes the filter state in
store. The BookList filter which is subscriebd to the store, recieves
filter as a props and based on the value of filter it renders the book
that have the category === filter.